+++ Source/HLEGraphics/uCodes/Ucode_GBI2.h
@@ -140,8 +140,8 @@ void DLParser_GBI2_MoveWord( MicroCodeCommand command )
 
 	case G_MW_FOG: // WIP, only works for a few games
 		{
-			f32 a = command.mw2.value >> 16;
-			f32 b = command.mw2.value & 0xFFFF;
+			f32 a = f32(command.mw2.value >> 16);
+			f32 b = f32(command.mw2.value & 0xFFFF);
 
 			//f32 min = b - a;
 			//f32 max = b + a;
